pyproject.toml
[tool.poetry]
name = "env-cschatbot"
version = "0.1.0"
description = "Env For Customer Service ChatBot"
authors = ["Mazhar <mazqoty.01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.3"
ipykernel = "^6.29.5"
pandas = "^2.2.3"
pinecone = "^5.4.0"
openai = "1.54.4"
python-dotenv = "^1.0.1"
scikit-learn = "^1.5.2"
easydict = "^1.13"
tiktoken = "^0.8.0"
pprintpp = "^0.4.0"
mlxtend = "0.23.1"
firebase-admin = "6.0.1"
google-cloud-storage = "2.18.2"
cloudinary = "^1.41.0"
tqdm = "^4.67.0"
rich = "^13.9.4"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
pyyaml = "^6.0.2"
colorama = "^0.4.6"
loguru = "^0.7.2"
pydantic = "^2.10.2"
pydantic-settings = "^2.6.1"
httpx = "0.27.2"


[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.3.1"
notebook = "^7.2.2"
ipywidgets = "^8.1.5"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
